requirements.txt
PyQt5>=5.15.9
openai>=1.3.0
requests>=2.31.0
psutil>=5.9.5
Pillow>=10.1.0
pytesseract>=0.3.10
deep-translator>=1.11.4
aiofiles>=23.2.1
chardet>=5.2.0
python-magic>=0.4.27
send2trash>=1.8.2
qrcode>=7.4.2
opencv-python>=4.8.1.78
numpy>=1.24.3
